Project refactor for template
