+++ Readme.md
@@ -1,0 +1,7 @@
+Gatsby Setup with Yarn Workspaces
+
+List of Packages
+
+1. Gatsby site setup with wordpress
+   to run the project use
+   -yarn workspace website develop
