+++ workspace/wordpress/gatsby-config.js
@@ -37,9 +37,9 @@ module.exports = {
       options: {
         typeName: `WPGraphQL`,
         fieldName: `wpgraphql`,
-        url:
-          // allows a fallback url if WPGRAPHQL_URL is not set in the env, this may be a local or remote WP instance.
-          process.env.GATSBY_WPGRAPHQL_URL,
+        url: 'http://localhost/web1/graphql',
+        // allows a fallback url if WPGRAPHQL_URL is not set in the env, this may be a local or remote WP instance.
+        // process.env.GATSBY_WPGRAPHQL_URL,
         schema: {
           //Prefixes all WP Types with "Wp" so "Post and allPost" become "WpPost and allWpPost".
           typePrefix: `Wp`,
