+++ workspace/wordpress/gatsby-node.js
@@ -1,8 +1,8 @@
 const path = require(`path`)
 
 exports.createPages = ({ graphql, actions }) => {
-    const { createPage } = actions
-    return graphql(`
+  const { createPage } = actions
+  return graphql(`
     {
         allWpPage {
             edges {
@@ -16,18 +16,18 @@ exports.createPages = ({ graphql, actions }) => {
     }
 
   `).then(result => {
-        console.log('post', result)
-        result.data.allWpPage.edges.forEach(({ node }) => {
-            createPage({
-                path: node.slug,
-                component: path.resolve(`./src/pages/page-2.js`),
-                context: {
-                    slug: node.slug,
-                    title: node.title,
-                    content: node.content
-                },
-            })
-        })
-    }
-    )
+    result.data.allWpPage.edges.forEach(({ node }) => {
+      console.log('node', node)
+      createPage({
+        path: node.slug,
+        component: path.resolve(`./src/templates/${node.slug}.js`),
+        context: {
+          slug: node.slug,
+          title: node.title,
+          content: node.content
+        },
+      })
+    })
+  }
+  )
 }
