+++ workspace/wordpress/src/pages/index.js
@@ -2,29 +2,17 @@ import React from "react"
 import { Link } from "gatsby"
 
 import Layout from "../components/layout"
-import Image from "../components/image"
 import SEO from "../components/seo"
-import Post from '../posts/index'
-const IndexPage = ({ data }) => (
-  <Layout>
-    <SEO title="Home" />
-    {data.allWpPost.edges.map(k => <Post post={k.node} key={k.node.title} />)}
-  </Layout>
-)
 
-export default IndexPage
+const SecondPage = ({ pageContext }) => {
 
-export const query = graphql`
-  query 
-    {
-        allWpPost {
-            edges {
-              node {
-                content
-                slug
-                title
-              }
-            }
-          }
-    
-  }`
+  const { title, content } = pageContext
+  console.log('context', pageContext)
+  return (
+    <Layout>
+      <SEO title="index" />
+      <h1>{title}</h1>
+      <div dangerouslySetInnerHTML={{ __html: content }} />
+    </Layout>)
+}
+export default SecondPage
