--- workspace/wordpress/src/pages/page-2.js
@@ -1,18 +1,0 @@
-import React from "react"
-import { Link } from "gatsby"
-
-import Layout from "../components/layout"
-import SEO from "../components/seo"
-
-const SecondPage = ({ pageContext }) => {
-
-  const { title, content } = pageContext
-  console.log('context', pageContext)
-  return (
-    <Layout>
-      <SEO title="Page two" />
-      <h1>{title}</h1>
-      <div dangerouslySetInnerHTML={{ __html: content }} />
-    </Layout>)
-}
-export default SecondPage
