+++ workspace/wordpress/src/posts/index.js
@@ -3,11 +3,13 @@ import Layout from "../components/layout"
 import { graphql } from "gatsby"
 
 export default ({ post }) => {
-    console.log('post', post)
     return (
         <div>
             <h1>{post.title}</h1>
             <div dangerouslySetInnerHTML={{ __html: post.content }} />
+            {
+                post.featuredImage && <img src={post.featuredImage.node.sourceUrl} alt={post.featuredImage.node.altText} />
+            }
         </div>
     )
 }
