+++ workspace/wordpress/src/templates/about.js
@@ -1,0 +1,18 @@
+import React from "react"
+import { Link } from "gatsby"
+
+import Layout from "../components/layout"
+import SEO from "../components/seo"
+
+const About = ({ pageContext }) => {
+
+    const { title, content } = pageContext
+    console.log('context', pageContext)
+    return (
+        <Layout>
+            <SEO title="Page two" />
+            <h1>{title}</h1>
+            <div dangerouslySetInnerHTML={{ __html: content }} />
+        </Layout>)
+}
+export default About
